# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: TheChickenMan B2B Chicken Marketplace API
  version: 1.0.0
  description: Authoritative RESTful API for TheChickenMan commercial poultry platform.
servers:
  - url: /api/v1
    description: Production & Local API Gateway
paths:
...
  /suppliers/profile:
    get:
      summary: Get current supplier profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Supplier profile }
    post:
      summary: Create or update supplier profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Supplier profile updated }
  /suppliers/kyc:
    post:
      summary: Submit supplier KYC details and documents
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: KYC submitted }
  /suppliers/locations:
    get:
      summary: List supplier locations
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: List of locations }
    post:
      summary: Add a new supplier location
      security: [{ BearerAuth: [] }]
      responses:
        '201': { description: Location created }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT